オプションウィンドウ追加・閉じるイベント追加
