+++ HookApp/Views/Option.xaml
@@ -1,0 +1,35 @@
+﻿<Window x:Class="HookApp.Views.Option"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:HookApp.Views"
+        mc:Ignorable="d"
+        Title="Option" Height="450" Width="800">
+
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="300px"></RowDefinition>
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="1*"></ColumnDefinition>
+            <ColumnDefinition Width="8*"></ColumnDefinition>
+            <ColumnDefinition Width="1*"></ColumnDefinition>
+        </Grid.ColumnDefinitions>
+
+        <GroupBox Grid.Row="0" Grid.Column="1" Header="スキンの選択" Margin="10px">
+
+            <ComboBox>
+                
+            </ComboBox>
+        </GroupBox>
+        <!--<StackPanel Grid.Row="0" Grid.Column="1" >
+            
+            <border
+            
+            <Label Content="aaa"></Label>
+        </StackPanel>-->
+
+    </Grid>
+
+</Window>
